+++ setup.sh
@@ -76,6 +76,7 @@ else
 fi
 
 echo "==> Directories -> ~/.config and \$HOME"
+link_dir redshift              "$HOME/.config/redshift"
 link_dir polybar              "$HOME/.config/polybar"
 link_dir newsboat             "$HOME/.newsboat"
 link_dir transmission-daemon  "$HOME/.config/transmission-daemon"